Babby secured cleanup.py

+++ somaris/kode/cleanup.py
@@ -8,7 +8,10 @@ patht = "tprotokoller/"
 pathpdf = "pdf/"
 pathtpdf = "temppdf/"
 
-os.system("rm " + pathf + "*")
-os.system("rm " + patht + "*")
-os.system("rm " + pathpdf + "*")
-os.system("rm " + pathtpdf + "*")
+if os.getcwd()[-13:] == "/somaris/kode":
+    os.system("rm " + pathf + "*")
+    os.system("rm " + patht + "*")
+    os.system("rm " + pathpdf + "*")
+    os.system("rm " + pathtpdf + "*")
+else:
+    print "Currently in wrong directory"
